refactor(doctor): extract telegram allowfrom scanning
